Ajuste Migrations
Ajuste Migrations

+++ src/HealthMed.Service.Medico/HealthMed.MedicoService.Domain/Entities/Medico.cs
@@ -7,7 +7,7 @@ public class Medico : EntityBase
     public required string Nome { get; set; }
     public required string Crm { get; set; }
     public required string Senha { get; set; }
-    public required int ValorConsulta { get; set; }
+    public required double ValorConsulta { get; set; }
     public bool Ativo { get; set; }
 
     public int EspecialidadeId { get; set; }
